# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Joy VoyagerApi module adds REST Api end points to Voyager with Passport and Swagger support https:\/\/github.com\/rxcod9\/joy-voyager-api.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/account.demo",
            "description": "Primary API Server"
        },
        {
            "url": "http:\/\/my-default-host.com",
            "description": "Another API Server"
        }
    ],
    "paths": {
        "\/api\/accountings": {
            "get": {
                "tags": [
                    "Accountings"
                ],
                "summary": "Get Accountings list",
                "operationId": "e7f1033252c0c75f8a0256df6c457c6b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "\/api\/accountings\/store": {
            "post": {
                "tags": [
                    "Accountings"
                ],
                "summary": "Accounting create",
                "operationId": "9a602eb32194724a2d61966f079f7e71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart\/form-data": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "string"
                                    },
                                    "money": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "\/api\/accountings\/{id}": {
            "get": {
                "tags": [
                    "Accountings"
                ],
                "summary": "Get accounting by show ID",
                "operationId": "851c830c654ffacce5906e1ed5e37772",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accountings"
                ],
                "summary": "Delete Task",
                "operationId": "5c78a84b45a0916668e7c1ca2a5f11b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "\/api\/change-accounting\/{acc}": {
            "put": {
                "tags": [
                    "Accountings"
                ],
                "summary": "Change accounting",
                "operationId": "4387661c7cdd9eeb0093ada96b9bcbf9",
                "parameters": [
                    {
                        "name": "acc",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart\/form-data": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "string"
                                    },
                                    "money": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        },
...
    }
}